+++ src/main/java/sh/ball/graph/blocks/BlockProcessor.java
@@ -1,0 +1,6 @@
+package sh.ball.graph.blocks;
+
+public interface BlockProcessor {
+
+  void process(double[] inputs, double[] outputs);
+}